+++ src/main/java/open/commons/spring/web/autoconfigure/configuration/GlobalServletConfiguration.java
@@ -27,13 +27,11 @@
 package open.commons.spring.web.autoconfigure.configuration;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import javax.annotation.Nonnull;
 import javax.servlet.http.HttpServletRequest;
@@ -57,12 +55,16 @@ import org.springframework.web.filter.OncePerRequestFilter;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import open.commons.core.utils.FunctionUtils;
+import open.commons.core.utils.StreamUtils;
 import open.commons.spring.web.config.ResourceConfiguration;
 import open.commons.spring.web.handler.DefaultGlobalInterceptor;
 import open.commons.spring.web.handler.HttpRequestProxyHeader;
 import open.commons.spring.web.handler.InterceptorIgnoreUrlProperties;
 import open.commons.spring.web.servlet.filter.AntPathRequest;
+import open.commons.spring.web.servlet.filter.RequestHeaderFilter;
 import open.commons.spring.web.servlet.filter.RequestThreadNameFilter;
+import open.commons.spring.web.servlet.filter.header.SharedHeader;
+import open.commons.spring.web.servlet.filter.header.SharedHeadersBuiltinProvider;
 
 /**
  * 
@@ -88,6 +90,9 @@ public class GlobalServletConfiguration {
     public static final String CONFIGURATION_BUILTIN_HTTP_REQUEST_PROXY_HEADER = "open.commons.spring.web.autoconfigure.configuration.GlobalServletConfiguration#CONFIGURATION_BUILTIN_HTTP_REQUEST_PROXY_HEADER";
     /** Proxy 서버를 통해서 전달되는 실제 클라이언트의 Http 요청 정보 경로 */
     public static final String PROPERTIES_HTTP_REQUEST_PROXY_HEADER = ResourceConfiguration.PROPERTIES_OPEN_COMMONS_SPRING_WEB_ROOT_PATH + ".proxy-header";
+
+    /** 외부에서 전달되는 로그 분기 정보를 처리하는 {@link SharedHeader} 제공 */
+    public static final String BEAN_QUALIFIER_PRIMARY_SHARED_HEADERS = "open.commons.spring.web.autoconfigure.configuration.GlobalServletConfiguration#PRIMARY_SHARED_HEADERS";
 
     private static final Logger logger = LoggerFactory.getLogger(GlobalServletConfiguration.class);
 
@@ -128,9 +133,42 @@ public class GlobalServletConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    @Order(Ordered.HIGHEST_PRECEDENCE)
+    @Order(RequestHeaderFilter.ORDER)
+    RequestHeaderFilter beanDefaultRequestHeaderFilter() {
+        return new RequestHeaderFilter();
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    @Order(RequestThreadNameFilter.ORDER)
     RequestThreadNameFilter beanDefaultRequestThreadNameFilter() {
         return new RequestThreadNameFilter();
+    }
+
+    /**
+     * {@link RequestThreadNameFilter} 이후에 위치시켜, {@link Thread} 이름 제어에 따르고 보안검증 이전에 Request 헤더 정보를 설정합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 19.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param filter
+     * @return
+     *
+     * @since 2025. 8. 19.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     */
+    @Bean
+    FilterRegistrationBean<RequestHeaderFilter> beanFilterRegistrationRequestHeaderFilter(RequestHeaderFilter filter) {
+        FilterRegistrationBean<RequestHeaderFilter> registration = new FilterRegistrationBean<>();
+        registration.setFilter(filter);
+        registration.setOrder(RequestHeaderFilter.ORDER); // 우선순위 상단
+        registration.addUrlPatterns("/*");
+        return registration;
     }
 
     /**
@@ -154,7 +192,7 @@ public class GlobalServletConfiguration {
     FilterRegistrationBean<RequestThreadNameFilter> beanFilterRegistrationThreadNamingFilter(RequestThreadNameFilter filter) {
         FilterRegistrationBean<RequestThreadNameFilter> registration = new FilterRegistrationBean<>();
         registration.setFilter(filter);
-        registration.setOrder(Ordered.HIGHEST_PRECEDENCE); // 우선순위 최상단
+        registration.setOrder(RequestThreadNameFilter.ORDER); // 우선순위 최상단
         registration.addUrlPatterns("/*");
         return registration;
     }
@@ -208,15 +246,10 @@ public class GlobalServletConfiguration {
     @Bean(name = BEAN_QUALIFIER_PRIMARY_INTERCEPTOR_IGNORE_URL_PATTERNS)
     @Primary
     Set<InterceptorIgnoreUrlProperties> beanPrimaryInterceptorIgnoreUrlPatterns( //
-            @NotNull @Nonnull Map<String, InterceptorIgnoreUrlProperties> singleConfigurations //
-            , @NotNull @Nonnull Map<String, List<InterceptorIgnoreUrlProperties>> multiConfigurations) {
+            @NotNull @Nonnull Map<String, InterceptorIgnoreUrlProperties> single //
+            , @NotNull @Nonnull Map<String, List<InterceptorIgnoreUrlProperties>> multi) {
 
-        List<InterceptorIgnoreUrlProperties> merged = Stream //
-                // 하나의 stream으로 병합
-                .of(singleConfigurations.values().stream() //
-                        , multiConfigurations.values().stream().flatMap(List::stream) //
-                ) //
-                .flatMap(s -> s).collect(Collectors.toList());
+        List<InterceptorIgnoreUrlProperties> merged = StreamUtils.toList(single, multi);
 
         // 중복 검증
         // key: FQCN 기반의 target 정보, value: 동일한 target 정보인 InterceptorIgnoreUrlProperties 객체들
@@ -235,19 +268,21 @@ public class GlobalServletConfiguration {
             }
         });
 
-        Map<String, InterceptorIgnoreUrlProperties> nego = new HashMap<>();
-        InterceptorIgnoreUrlProperties o = null;
-        for (InterceptorIgnoreUrlProperties p : merged) {
-            o = nego.get(p.getTarget());
-            if (o != null) {
-                o.addExcludePathPatterns(p.getExcludePathPatterns());
-                o.addIncludePathPatterns(p.getIncludePathPatterns());
-            } else {
-                nego.put(p.getTarget(), p);
-            }
-        }
+        Set<InterceptorIgnoreUrlProperties> result = StreamUtils.toSet(merged, InterceptorIgnoreUrlProperties::getTarget //
+                , // 설정 객체 데이터 변조 방지를 위해 새로운 객체 생성
+                p -> {
+                    InterceptorIgnoreUrlProperties n = new InterceptorIgnoreUrlProperties();
+                    n.setTarget(p.getTarget());
+                    n.setExcludePathPatterns(p.getExcludePathPatterns());
+                    n.setIncludePathPatterns(p.getIncludePathPatterns());
+                    return n;
+                }, (old, income) -> {
+                    old.addExcludePathPatterns(income.getExcludePathPatterns());
+                    old.addIncludePathPatterns(income.getIncludePathPatterns());
+                    return old;
+                });
 
-        return nego.values().stream().collect(Collectors.toSet());
+        return result;
     }
 
     @Bean(name = BEAN_QUALIFIER_PRIMARY_ONCE_PER_REQUEST_SHOULD_NOT_PATTERNS)
@@ -255,11 +290,16 @@ public class GlobalServletConfiguration {
     List<AntPathRequest> beanPrimaryOncePerRequestShouldNotFilters( //
             @NotNull @Nonnull Map<String, AntPathRequest> single//
             , @NotNull @Nonnull Map<String, List<AntPathRequest>> multi) {
-        return Stream //
-                .of(single.values().stream() //
-                        , multi.values().stream().flatMap(List::stream) //
-                ) //
-                .flatMap(s -> s).collect(Collectors.toList());
+        return StreamUtils.toList(single, multi);
+    }
+
+    @Bean(name = BEAN_QUALIFIER_PRIMARY_SHARED_HEADERS)
+    @Primary
+    List<SharedHeader> beanPrimarySharedHeaders( //
+            @NotNull @Nonnull Map<String, SharedHeader> single //
+            , @NotNull @Nonnull Map<String, List<SharedHeader>> multi //
+    ) {
+        return StreamUtils.toList(single, multi);
     }
 
     /**
@@ -402,5 +442,26 @@ public class GlobalServletConfiguration {
         paths.add(new AntPathRequest("/swagger-ui/**", HttpMethod.GET));
 
         return paths;
+    }
+
+    /**
+     * 내부에서 정의된 {@link SharedHeader} 정보를 제공합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 20.		박준홍			최초 작성
+     * </pre>
+     *
+     * @return
+     *
+     * @since 2025. 8. 20.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     */
+    @Bean
+    List<SharedHeader> configBuiltinSharedHeaders() {
+        return SharedHeadersBuiltinProvider.load();
     }
 }