+++ src/main/java/open/commons/spring/web/servlet/filter/AntPathRequest.java
@@ -27,6 +27,7 @@
 package open.commons.spring.web.servlet.filter;
 
 import javax.annotation.Nonnull;
+import javax.servlet.Filter;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.http.HttpMethod;
@@ -52,6 +53,11 @@ public class AntPathRequest {
     private HttpMethod httpMethod;
     /** URL 패턴 대/소문자 구분 여부 */
     private boolean caseSensitive = false;
+
+    /** 패턴을 적용하는 방식 */
+    private Scheme targetType = Scheme.CLASS;
+    /** {@link Filter} 구현 클래스. */
+    private Class<?> filterClass;
 
     /**
      * <br>
@@ -82,13 +88,14 @@ public class AntPathRequest {
      * </pre>
      *
      * @param pattern
+     *            URL 패턴
      *
      * @since 2025. 8. 4.
      * @version 0.8.0
      * @author parkjunhong77@gmail.com
      */
-    public AntPathRequest(@NotNull String pattern) {
-        this(pattern, null, false);
+    public AntPathRequest(@NotNull @Nonnull String pattern) {
+        this(pattern, null, false, null, null);
     }
 
     /**
@@ -102,14 +109,16 @@ public class AntPathRequest {
      * </pre>
      *
      * @param pattern
+     *            URL 패턴
      * @param httpMethod
+     *            요청 {@link HttpMethod}
      *
      * @since 2025. 8. 4.
      * @version 0.8.0
      * @author parkjunhong77@gmail.com
      */
-    public AntPathRequest(@NotNull String pattern, HttpMethod httpMethod) {
-        this(pattern, httpMethod, false);
+    public AntPathRequest(@NotNull @Nonnull String pattern, HttpMethod httpMethod) {
+        this(pattern, httpMethod, false, null, null);
     }
 
     /**
@@ -123,17 +132,100 @@ public class AntPathRequest {
      * </pre>
      *
      * @param pattern
+     *            URL 패턴
      * @param httpMethod
+     *            요청 {@link HttpMethod}
      * @param caseSensitive
+     *            URL 대소문자 비교여부
      *
      * @since 2025. 8. 4.
      * @version 0.8.0
      * @author parkjunhong77@gmail.com
      */
-    public AntPathRequest(@NotNull String pattern, HttpMethod httpMethod, boolean caseSensitive) {
+    public AntPathRequest(@NotNull @Nonnull String pattern, HttpMethod httpMethod, boolean caseSensitive) {
+        this(pattern, httpMethod, caseSensitive, null, null);
+    }
+
+    /**
+     * 
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 20.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param pattern
+     *            URL 패턴
+     * @param httpMethod
+     *            요청 {@link HttpMethod}
+     * @param caseSensitive
+     *            URL 대소문자 비교여부
+     * @param targetType
+     *            비교 방식
+     * @param filterClass
+     *            {@link Filter} 구현 클래스
+     * @since 2025. 8. 20.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     */
+    public AntPathRequest(@NotNull @Nonnull String pattern, HttpMethod httpMethod, boolean caseSensitive, Scheme targetType, Class<? extends Filter> filterClass) {
         this.pattern = pattern;
         this.httpMethod = httpMethod;
         this.caseSensitive = caseSensitive;
+        this.targetType = targetType;
+        this.filterClass = filterClass;
+    }
+
+    /**
+     * 
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 20.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param pattern
+     *            URL 패턴
+     * @param targetType
+     *            비교 방식
+     * @param filterClass
+     *            {@link Filter} 구현 클래스
+     *
+     * @since 2025. 8. 20.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     */
+    public AntPathRequest(@NotNull @Nonnull String pattern, Scheme targetType, Class<? extends Filter> filterClass) {
+        this(pattern, null, false, targetType, filterClass);
+    }
+
+    /**
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 19.		박준홍			최초 작성
+     * </pre>
+     * 
+     * @return the filterClass
+     *
+     * @since 2025. 8. 19.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     *
+     * @see #filterClass
+     */
+
+    public Class<?> getFilterClass() {
+        return filterClass;
     }
 
     /**
@@ -196,6 +288,29 @@ public class AntPathRequest {
      * [개정이력]
      *      날짜    	| 작성자	|	내용
      * ------------------------------------------
+     * 2025. 8. 19.		박준홍			최초 작성
+     * </pre>
+     * 
+     * @return the targetType
+     *
+     * @since 2025. 8. 19.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     *
+     * @see #targetType
+     */
+
+    public Scheme getTargetType() {
+        return targetType;
+    }
+
+    /**
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
      * 2025. 8. 4.		박준홍			최초 작성
      * </pre>
      * 
@@ -209,6 +324,40 @@ public class AntPathRequest {
      */
     public boolean isCaseSensitive() {
         return caseSensitive;
+    }
+
+    /**
+     * 이 객체가 {@link Filter}를 구현한 클래스에 적용가능한지 여부를 제공합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 19.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param <T>
+     * @param filterClass
+     *            {@link Filter} 구현 클래스.
+     * @return
+     *
+     * @since 2025. 8. 19.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     */
+    public <T extends Filter> boolean matches(@Nonnull Class<T> filterClass) {
+        if (this.filterClass == null) {
+            return true;
+        } else {
+            switch (this.targetType) {
+                case CLASS:
+                    return this.filterClass.equals(filterClass);
+                case INSTANCE:
+                    return this.filterClass.isAssignableFrom(filterClass);
+                default:
+                    return false;
+            }
+        }
     }
 
     /**
@@ -232,6 +381,29 @@ public class AntPathRequest {
      */
     public void setCaseSensitive(boolean caseSensitive) {
         this.caseSensitive = caseSensitive;
+    }
+
+    /**
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 19.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param filterClass
+     *            the filterClass to set
+     *
+     * @since 2025. 8. 19.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     *
+     * @see #filterClass
+     */
+    public void setFilterClass(Class<?> filterClass) {
+        this.filterClass = filterClass;
     }
 
     /**
@@ -284,8 +456,33 @@ public class AntPathRequest {
     }
 
     /**
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 8. 19.		박준홍			최초 작성
+     * </pre>
      *
-     * @since 2025. 8. 4.
+     * @param targetType
+     *            the targetType to set
+     *
+     * @since 2025. 8. 19.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     *
+     * @see #targetType
+     */
+    public void setTargetType(Scheme targetType) {
+        if (targetType != null) {
+            this.targetType = targetType;
+        }
+    }
+
+    /**
+     *
+     * @since 2025. 8. 19.
      * @version 0.8.0
      * @author parkjunhong77@gmail.com
      *
@@ -300,8 +497,18 @@ public class AntPathRequest {
         builder.append(httpMethod);
         builder.append(", caseSensitive=");
         builder.append(caseSensitive);
+        builder.append(", targetType=");
+        builder.append(targetType);
+        builder.append(", filterClass=");
+        builder.append(filterClass);
         builder.append("]");
         return builder.toString();
     }
 
+    public static enum Scheme {
+        /** 일치 */
+        CLASS,
+        /** 구현 클래스 */
+        INSTANCE
+    }
 }